# Optimize BufferedSubStream.ReadByte